statusをnumberからintegerに変更

+++ lib/models/swagger/v1/model/task_schema.dart
@@ -7,7 +7,7 @@ class TaskSchema {
 
   String? name;
 
-  double? status;
+  int? status;
 
   String? createdAt;
